[VUMM-40] move all exceptions classes to common module for simplicity
