+++ backend/src/main/java/ai/verta/modeldb/datasetVersion/DatasetVersionDAO.java
@@ -3,9 +3,9 @@ package ai.verta.modeldb.datasetVersion;
 import ai.verta.common.KeyValue;
 import ai.verta.modeldb.DatasetVersion;
 import ai.verta.modeldb.FindDatasetVersions;
+import ai.verta.modeldb.common.exceptions.PermissionDeniedException;
 import ai.verta.modeldb.dataset.DatasetDAO;
 import ai.verta.modeldb.dto.DatasetVersionDTO;
-import ai.verta.modeldb.exceptions.PermissionDeniedException;
 import ai.verta.uac.UserInfo;
 import java.util.List;
 import org.hibernate.Session;
