--- backend/src/main/java/ai/verta/modeldb/exceptions/PermissionDeniedException.java
@@ -1,10 +1,0 @@
-package ai.verta.modeldb.exceptions;
-
-import ai.verta.modeldb.common.exceptions.ModelDBException;
-import com.google.rpc.Code;
-
-public class PermissionDeniedException extends ModelDBException {
-  public PermissionDeniedException(String message) {
-    super(message, Code.PERMISSION_DENIED);
-  }
-}
