+++ backend/src/main/java/ai/verta/modeldb/experimentRun/ExperimentRunDAO.java
@@ -6,8 +6,8 @@ import ai.verta.common.KeyValue;
 import ai.verta.modeldb.*;
 import ai.verta.modeldb.CommitArtifactPart.Response;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
+import ai.verta.modeldb.common.exceptions.PermissionDeniedException;
 import ai.verta.modeldb.dto.ExperimentRunPaginationDTO;
-import ai.verta.modeldb.exceptions.PermissionDeniedException;
 import ai.verta.modeldb.versioning.CommitFunction;
 import ai.verta.modeldb.versioning.EnvironmentBlob;
 import ai.verta.modeldb.versioning.RepositoryFunction;
