+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/CodeVersionHandler.java
@@ -2,10 +2,10 @@ package ai.verta.modeldb.experimentRun.subtypes;
 
 import ai.verta.common.CodeVersion;
 import ai.verta.modeldb.ModelDBConstants;
+import ai.verta.modeldb.common.exceptions.AlreadyExistsException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.entities.CodeVersionEntity;
-import ai.verta.modeldb.exceptions.AlreadyExistsException;
 import ai.verta.modeldb.utils.ModelDBHibernateUtil;
 import ai.verta.modeldb.utils.RdbmsUtils;
 import java.util.AbstractMap;
