+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/HyperparameterPredicatesHandler.java
@@ -3,9 +3,9 @@ package ai.verta.modeldb.experimentRun.subtypes;
 import ai.verta.common.KeyValueQuery;
 import ai.verta.common.OperatorEnum;
 import ai.verta.modeldb.common.CommonUtils;
+import ai.verta.modeldb.common.exceptions.UnimplementedException;
 import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.query.QueryFilterContext;
-import ai.verta.modeldb.exceptions.UnimplementedException;
 import java.util.Date;
 
 public class HyperparameterPredicatesHandler extends PredicateHandlerUtils {
