+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/PredicatesHandler.java
@@ -10,9 +10,9 @@ import ai.verta.modeldb.common.EnumerateList;
 import ai.verta.modeldb.common.dto.UserInfoPaginationDTO;
 import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
+import ai.verta.modeldb.common.exceptions.UnimplementedException;
 import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.query.QueryFilterContext;
-import ai.verta.modeldb.exceptions.UnimplementedException;
 import ai.verta.modeldb.utils.UACApisUtil;
 import ai.verta.uac.GetResourcesResponseItem;
 import ai.verta.uac.UserInfo;
